--- a/pkg/env/sensor.go
+++ b/pkg/env/sensor.go
@@ -93,4 +93,14 @@ var (
 	// ResponsesChannelBufferSize defines how many messages to central are we buffering before dropping messages
 	// Setting this variable to zero will disable this feature.
 	ResponsesChannelBufferSize = RegisterIntegerSetting("ROX_RESPONSES_CHANNEL_BUFFER_SIZE", 0)
+
+	// EnrichmentPurgerTickerMaxAge controls the max age of collector updates (network flows & container endpoints)
+	// for keeping them in  Sensor's memory. Entries that has not been enriched (due to a bug or error)
+	// will stay in Sensors memory until restart. Purger cleans all those entries based on rules.
+	// The max-age is a rule of last resort (when all other rules do not apply) and is used to protect Sensor from OOM kills.
+	// Set to zero to not purge based on max-age (other purger rules will be executed).
+	EnrichmentPurgerTickerMaxAge = registerDurationSetting("ROX_ENRICHMENT_PURGER_MAX_AGE", 4*time.Hour, WithDurationZeroAllowed())
+	// EnrichmentPurgerTickerCycle controls how frequently purger is run to check for collector updates
+	// (network flows & container endpoints) that stuck in Sensor's memory. Set to zero to completely disable the purger.
+	EnrichmentPurgerTickerCycle = registerDurationSetting("ROX_ENRICHMENT_PURGER_UPDATE_CYCLE", 30*time.Minute, WithDurationZeroAllowed())
 )