--- a/sensor/common/clusterentities/store.go
+++ b/sensor/common/clusterentities/store.go
@@ -265,10 +265,10 @@ func (e *Store) LookupByEndpoint(endpoint net.NumericEndpoint) []LookupResult {
 }
 
 // LookupByContainerID retrieves the deployment ID by a container ID.
-func (e *Store) LookupByContainerID(containerID string) (result ContainerMetadata, found bool) {
-	result, found, _ = e.containerIDsStore.lookupByContainer(containerID)
+func (e *Store) LookupByContainerID(containerID string) (metadata ContainerMetadata, found bool, isHistorical bool) {
+	metadata, found, isHistorical = e.containerIDsStore.lookupByContainer(containerID)
 	e.track("LookupByContainerID(%s): found=%t", containerID, found)
-	return result, found
+	return metadata, found, isHistorical
 }
 
 // RegisterPublicIPsListener registers a listener that listens on changes to the set of public IP addresses.