--- a/sensor/common/networkflow/manager/interfaces.go
+++ b/sensor/common/networkflow/manager/interfaces.go
@@ -9,7 +9,7 @@ import (
 //
 //go:generate mockgen-wrapper
 type EntityStore interface {
-	LookupByContainerID(string) (clusterentities.ContainerMetadata, bool)
+	LookupByContainerID(string) (metadata clusterentities.ContainerMetadata, found bool, isHistorical bool)
 	LookupByEndpoint(net.NumericEndpoint) []clusterentities.LookupResult
 	RegisterPublicIPsListener(clusterentities.PublicIPsListener) bool
 	UnregisterPublicIPsListener(clusterentities.PublicIPsListener) bool