--- a/sensor/common/networkflow/manager/manager_impl.go
+++ b/sensor/common/networkflow/manager/manager_impl.go
@@ -379,7 +379,7 @@ func (m *networkFlowManager) Start() error {
 	go m.publicIPs.Run(m.stopper.LowLevel().GetStopRequestSignal(), m.clusterEntities)
 	if m.purger != nil {
 		if err := m.purger.Start(); err != nil {
-			return errors.Wrap(err, "unable to run network flow purger")
+			return errors.Wrap(err, "unable to start network flow purger")
 		}
 	}
 	return nil