--- a/sensor/common/networkflow/manager/purger.go
+++ b/sensor/common/networkflow/manager/purger.go
@@ -85,9 +85,6 @@ func (p *NetworkFlowPurger) Stop(_ error) {
 
 // nonZeroPurgerCycle delivers a non-zero duration to be used in timers (they panic when set with 0 duration)
 func nonZeroPurgerCycle() time.Duration {
-	// How often purger should run. The Purger removes old endpoints from activeEndpoints slice.
-	// This is important for cases when Collector or the orchestrator never reports a given endpoint
-	// as deleted, because there is no other mechanism that would remove an endpoint from memory.
 	purgerCycleSetting := env.EnrichmentPurgerTickerCycle.DurationSetting()
 	if purgerCycleSetting > 0 {
 		return purgerCycleSetting